vision-critique/pyproject.toml
[project]
name = "vision-critique"
version = "0.1.0"
description = "Vision-based UI quality critique tool for coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AutoReport Team" }
]

dependencies = [
    "playwright>=1.40.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "cssutils>=2.9.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.scripts]
vision-critique = "vision_critique.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
